requirements.txt
spotipy
requests
mysql-connector-python
mysql.connector